requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
openpyxl>=3.1.0
fredapi>=2.1.0
python-dotenv>=1.0.0